Pequenas correções

+++ class/lengthData.cpp
@@ -1,9 +1,9 @@
 #ifndef LENGTHDATA
 #define LENGTHDATA
 
-typedef unsigned short int vertexLabelType;
-typedef unsigned short int edgeLabelType;
-typedef unsigned short int edgesTotalLabelType;
-typedef unsigned short int vertexesTotalLabelType;
+typedef unsigned int vertexLabelType;
+typedef unsigned int edgeLabelType;
+typedef unsigned int edgesTotalLabelType;
+typedef unsigned int vertexesTotalLabelType;
 
 #endif
